+++ knp/tests/core/projection_test.cpp
@@ -20,21 +20,21 @@ using SynapseGenerator = DeltaProjection::SynapseGenerator;
 
 SynapseGenerator make_dense_generator(std::pair<size_t, size_t> pop_sizes, SynapseParameters default_params)
 {
-    SynapseGenerator generator = [=](size_t index) -> std::optional<Synapse>
+    SynapseGenerator generator = [=](uint32_t index) -> std::optional<Synapse>
     {
-        const size_t id_from = index / pop_sizes.second;
-        const size_t id_to = index % pop_sizes.second;
+        const uint32_t id_from = index / pop_sizes.second;
+        const uint32_t id_to = index % pop_sizes.second;
         return Synapse{default_params, id_from, id_to};
     };
     return generator;
 }
 
 
-SynapseGenerator make_cyclic_generator(std::pair<size_t, size_t> pop_sizes, SynapseParameters default_params)
+SynapseGenerator make_cyclic_generator(std::pair<uint32_t, uint32_t> pop_sizes, SynapseParameters default_params)
 {
-    SynapseGenerator generator = [=](size_t index) -> std::optional<Synapse>
+    SynapseGenerator generator = [=](uint32_t index) -> std::optional<Synapse>
     {
-        const size_t index_from = index % pop_sizes.first;
+        const uint32_t index_from = index % pop_sizes.first;
         return Synapse{default_params, index_from, (index_from + index / pop_sizes.second) % pop_sizes.second};
     };
     return generator;
@@ -49,10 +49,10 @@ TEST(ProjectionSuite, Generation)
     const float weight_constant = 0.0001;
 
     // Creating a dense projection with random weights ( -0.12 to +0.17 ) and delay (1 to 5)
-    SynapseGenerator generator = [&](size_t iter) -> std::optional<Synapse>
+    SynapseGenerator generator = [&](uint32_t iter) -> std::optional<Synapse>
     {
-        const size_t id_from = iter / postsynaptic_size;
-        const size_t id_to = iter % postsynaptic_size;
+        const uint32_t id_from = iter / postsynaptic_size;
+        const uint32_t id_to = iter % postsynaptic_size;
         const SynapseParameters params(weight_constant * iter, iter / 100 + 1);
         return std::optional(Synapse{params, id_from, id_to});
     };
@@ -68,15 +68,15 @@ TEST(ProjectionSuite, Generation)
 TEST(ProjectionSuite, SynapseAddition)
 {
     // TODO: add_synapses, remove_presynaptic_neuron, remove_postsynaptic_neuron, disconnect,
-    const size_t presynaptic_size = 1000;
-    const size_t postsynaptic_size = presynaptic_size;
-    const size_t neuron_index = 10;  // A specific neuron
+    const uint32_t presynaptic_size = 1000;
+    const uint32_t postsynaptic_size = presynaptic_size;
+    const uint32_t neuron_index = 10;  // A specific neuron
 
     // Create an empty projection
     DeltaProjection projection(knc::UID{}, knc::UID{});
 
     // Add synapses to an empty projection using a generator
-    SynapseGenerator generator1 = [](size_t index) { return Synapse{{0.0F, 1}, index, index}; };
+    SynapseGenerator generator1 = [](uint32_t index) { return Synapse{{0.0F, 1}, index, index}; };
     projection.add_synapses(presynaptic_size, generator1);  // Add 1-to-1 synapses
     ASSERT_EQ(projection.size(), presynaptic_size);
     size_t count = projection.add_synapses(
@@ -84,7 +84,7 @@ TEST(ProjectionSuite, SynapseAddition)
     ASSERT_EQ(count, 1);
 
     // Add synapses to nonempty projection using a generator
-    SynapseGenerator generator2 = [](size_t index) {
+    SynapseGenerator generator2 = [](uint32_t index) {
         return Synapse{{0.1F, 2}, index, (index + 1) % postsynaptic_size};
     };
     count = projection.add_synapses(presynaptic_size, generator2);  // Add synapses from pre neuron N to post neuron N+1
